requirements.txt
flask==3.1.0
flask-cors==6.0.0
openai==1.78.0
python-dotenv==1.1.0
gunicorn
beautifulsoup4
requests
mutagen
Pillow
PyPDF2